requirements.txt
py-cord>=2.4.1,<3.0.0
python-dotenv>=1.0.0,<2.0.0
openai>=0.27.4,<1.0.0
